Bash file for running flink project
The flink job can only run when you output its contents to a textfile as a data sink so just change the path of the text file to the path of an already existing text file.The recompile the package but typing (mvn clean package) in the root of the location streaming folder. You might have to change the port in the Location.py script to match the one in StreamingJob.java...I had a few blocked ports so I had to change it. The target folder from the location streaming folder must be copied and pasted in the flink folder (the same folder as bin ).The run the start-local.sh to see the job on the browser.